+++ backend/canisters/local_user_index/api/src/lib.rs
@@ -2,11 +2,11 @@ use candid::{CandidType, Principal};
 use serde::{Deserialize, Serialize};
 use types::nns::CryptoAmount;
 use types::{
-    is_default, AutonomousConfig, BotDefinition, BotInstallationLocation, CanisterId, ChannelLatestMessageIndex, ChatId,
-    ChitEarnedReason, CommunityId, Cryptocurrency, DiamondMembershipPlanDuration, MessageContent, MessageContentInitial,
-    MessageId, MessageIndex, NotifyChit, PhoneNumber, ReferralType, SlashCommandSchema, SuspensionDuration, TimestampMillis,
-    UniquePersonProof, UpdateUserPrincipalArgs, User, UserCanisterStreakInsuranceClaim, UserCanisterStreakInsurancePayment,
-    UserId, UserType,
+    is_default, AutonomousConfig, BotCommandDefinition, BotDefinition, BotInstallationLocation, CanisterId,
+    ChannelLatestMessageIndex, ChatId, ChitEarnedReason, CommunityId, Cryptocurrency, DiamondMembershipPlanDuration,
+    MessageContent, MessageContentInitial, MessageId, MessageIndex, NotifyChit, PhoneNumber, ReferralType, SuspensionDuration,
+    TimestampMillis, UniquePersonProof, UpdateUserPrincipalArgs, User, UserCanisterStreakInsuranceClaim,
+    UserCanisterStreakInsurancePayment, UserId, UserType,
 };
 
 mod lifecycle;
@@ -90,7 +90,7 @@ pub struct BotRegistered {
     pub user_id: UserId,
     pub user_principal: Principal,
     pub name: String,
-    pub commands: Vec<SlashCommandSchema>,
+    pub commands: Vec<BotCommandDefinition>,
     pub autonomous_config: Option<AutonomousConfig>,
 }