+++ backend/libraries/types/src/c2c_bot_api_key.rs
@@ -1,4 +1,11 @@
+use crate::UserId;
 use serde::{Deserialize, Serialize};
+
+#[derive(Serialize, Deserialize, Debug)]
+pub struct Args {
+    pub bot_id: UserId,
+    pub initiator: UserId,
+}
 
 #[derive(Serialize, Deserialize, Debug)]
 pub enum Response {